modified min size to be a int

+++ run_pol_lda_mallet.py
@@ -33,5 +33,5 @@ filter_array = np.logical_and(stat_df.language == 'en',
 
 print(f"{len(filter_array)} is limiting to {sum(filter_array)}")
 nlp.preprocess(load_existing=True, filter_loaded=filter_array)
-nlp.create_bag_of_words(filter_extremes=True, min_df=(0.001*sum(filter_array)), max_df=0.3)
+nlp.create_bag_of_words(filter_extremes=True, min_df=(int(0.001*sum(filter_array))), max_df=0.3)
 nlp.search_best_model_mallet(topic_list=list(range(5, 100, 5)))
